+++ EventPassBE/EventPass.Application/Commands/Carts/Create/CreateCartCommandHandler.cs
@@ -25,7 +25,7 @@ namespace EventPass.Application.Commands.Carts.Create
             return new ResponseCartDto
             {
                 Id = response.Id,
-                userId = response.UserID
+                userId = (int)response.UserID
             };
         }
     }
