+++ EventPassBE/EventPass.Application/Commands/SponsorEvents/Update/UpdateSponsorEventCommandHandler.cs
@@ -28,8 +28,8 @@ namespace EventPass.Application.Commands.SponsorEvents.Update
                 Id = response.Id,
                 tier = response.Tier,
                 amountSponsored = response.AmountSponsored,
-                sponsorId = response.SponsorID,
-                eventId = response.EventID
+                sponsorId = (int)response.SponsorID,
+                eventId = (int)response.EventID
             };
         }
     }
