+++ EventPassBE/EventPass.Application/Queries/Orders/GetAll/GetAllOrdersQueryHandler.cs
@@ -28,7 +28,7 @@ namespace EventPass.Application.Queries.Orders.GetAll
                 BillingAdress = order.BillingAdress,
                 ShippingAdress = order.ShippingAdress,
                 TransactionNumber = order.TransactionNumber,
-                UserId = order.UserID                
+                UserId = (int)order.UserID                
             });
         }
     }
