+++ EventPassBE/EventPass.Application/Queries/Reviews/GetAll/GetAllReviewsQueryHandler.cs
@@ -22,8 +22,8 @@ namespace EventPass.Application.Handlers.Reviews.GetAll
             {
                 Id = r.Id,
                 Rating = r.Rating,
-                PerformerID = r.PerformerID,
-                UserID = r.UserID,
+                PerformerID = (int)r.PerformerID,
+                UserID = (int)r.UserID,
                 UserName = r.User.Name,
                 PerformerName = r.Performer.Name
             });
