+++ EventPassBE/EventPass.Application/Queries/Reviews/GetByPerformer/GetReviewsByPerformerQueryHandler.cs
@@ -23,8 +23,8 @@ namespace EventPass.Application.Handlers.Reviews.GetByPerformer
             {
                 Id = r.Id,
                 Rating = r.Rating,
-                PerformerID = r.PerformerID,
-                UserID = r.UserID,
+                PerformerID = (int)r.PerformerID,
+                UserID = (int)r.UserID,
                 UserName = r.User.Name,
                 PerformerName = r.Performer.Name
             });
